Refactor project index parser error formatting
